fix: handle unsupported cached textures in USDZ export
